requirements.txt
altair
numpy
pandas
pydeck
streamlit
superpowered-sdk
streamlit-scrollable-textbox
psycopg2-binary
sqlalchemy
extra-streamlit-components
streamlit_js_eval